// hogwarts/src/constants/enum/constants.enum.ts
export enum ConstantName {
  isLightOn = 'isLightOn',
  veritaserum = 'veritaserum',
  owlAddress = 'owlAdress',
  owlTimeout = 'owlTimeout',
  owlAddressMinistryResponse = 'owlAddressMinistryResponse',
  owlPhoenixOrderResponse = 'owlPhoenixOrderResponse',
}

export enum VeritaserumType {
  everything = 'EVERYTHING',
  half = 'HALF',
  nothing = 'NOTHING',
}

export const owlAddressMinistry = 'http://magic-ministry.com/';
export const owlAddressOrder = 'http://phoenix-order.com/';